requirements.txt
slpp
pyyaml
pytest